Split the preset agent module by concern
agent.py had grown to 1591 lines mixing six concerns. Split it, moving every
symbol verbatim:

- session.py: on-disk session state, loaders, ownership claims, liveness
- workspace.py: the agent working directory, aliasing, wrappers, skills
- tail.py: offset-persistent tailers over stream and record files
- redaction.py: secret redaction shared by the tailers and verify
- agent.py: the Claude subprocess lifecycle (600 lines)

Imports follow one direction: session <- workspace <- (tail, redaction) <-
agent. No behavior change; all symbol bodies are byte-identical to before.
